extract widgets from create_screen
